Add files via upload
The names refer to the inheritance structure. For example, "plat" may be instantiated under the base class "region". 

pair_plat are source code for the aggregation force between platelet particles. 

region_ccylinder are source code for a torus-shaped geometric object during simulation. Some of the methods are left blank with only error output, as they are not needed.